Subfont workaround.
Subfont does not automatically use the font subsets for all font-family declarations.
This probably stems from styled-components v4 createGlobalStyle and may be resolved with a future update to styled-components or gatsby-plugin-styled-components.

As a workaround, the __subset classes can be specified manually in globalStyle.js and hidden divs can ensure that subfont includes all required letters in the subsets.